+++ src/main/java/gov/usgs/volcanoes/swarm/wave/WaveViewSettings.java
@@ -45,35 +45,51 @@ public class WaveViewSettings {
     }
   }
 
-  public boolean filterOn;
-  public boolean zeroPhaseShift;
-  public boolean autoScaleAmp;
-  public boolean autoScaleAmpMemory;
-  public boolean autoScalePower;
-  public boolean autoScalePowerMemory;
-  public boolean useUnits;
-  public boolean logPower;
-  public boolean logFreq;
-  public boolean removeBias;
-
-  public double maxAmp;
-  public double minAmp;
-  public double minPower;
-  public double maxPower;
-  public double minFreq;
-  public double maxFreq;
-  public double spectrogramOverlap;
-  public double binSize;
-
-  public int nfft;
 
   public WaveViewPanel view;
   public WaveViewSettingsToolbar toolbar;
   public ViewType viewType;
-  public Butterworth filter;
   
+  // wave settings
+  public boolean autoScaleAmp;
+  public boolean autoScaleAmpMemory;
+  public double waveMaxAmp;
+  public double waveMinAmp;
+  public boolean removeBias;
+  public boolean useUnits;
+  
+  // spectra settings
+  public boolean spectraLogPower;
+  public boolean spectraLogFreq;
+  public boolean autoScaleSpectraPower;
+  public boolean autoScaleSpectraPowerMemory;
+  public double spectraMinPower;
+  public double spectraMaxPower;
+  public double spectraMinFreq;
+  public double spectraMaxFreq;
+  
+  // spectrogram settings
+  public boolean spectrogramLogPower;
+  public boolean autoScaleSpectrogramPower;
+  public boolean autoScaleSpectrogramPowerMemory;
+  public double spectrogramMinPower;
+  public double spectrogramMaxPower;
+  public double spectrogramMinFreq;
+  public double spectrogramMaxFreq;
+  public double spectrogramOverlap;
+  public double binSize;
+  public int nfft;
+  
+  // particle motion settings
   public boolean useAlternateOrientationCode = false;
   public String alternateOrientationCode = "Z12";
+  
+  // filter settings
+  public boolean filterOn;
+  public Butterworth filter;
+  public boolean zeroPhaseShift;
+  
+  // pick mode and tag mode settings
   public boolean pickEnabled = false;
   public boolean tagEnabled = false;
 
@@ -81,29 +97,42 @@ public class WaveViewSettings {
 
   static {
     DEFAULT_WAVE_VIEW_SETTINGS = new WaveViewSettings();
+    // view option
     DEFAULT_WAVE_VIEW_SETTINGS.viewType = ViewType.WAVE;
+    // wave options
     DEFAULT_WAVE_VIEW_SETTINGS.removeBias = true;
     DEFAULT_WAVE_VIEW_SETTINGS.autoScaleAmp = true;
     DEFAULT_WAVE_VIEW_SETTINGS.autoScaleAmpMemory = true;
-    DEFAULT_WAVE_VIEW_SETTINGS.maxAmp = 1000;
-    DEFAULT_WAVE_VIEW_SETTINGS.minAmp = -1000;
-    DEFAULT_WAVE_VIEW_SETTINGS.autoScalePower = false;
-    DEFAULT_WAVE_VIEW_SETTINGS.autoScalePowerMemory = true;
-    DEFAULT_WAVE_VIEW_SETTINGS.minPower = 20;
-    DEFAULT_WAVE_VIEW_SETTINGS.maxPower = 120;
+    DEFAULT_WAVE_VIEW_SETTINGS.waveMaxAmp = 1000;
+    DEFAULT_WAVE_VIEW_SETTINGS.waveMinAmp = -1000;
     DEFAULT_WAVE_VIEW_SETTINGS.useUnits = true;
-    DEFAULT_WAVE_VIEW_SETTINGS.logPower = true;
-    DEFAULT_WAVE_VIEW_SETTINGS.logFreq = true;
+    // spectra options
+    DEFAULT_WAVE_VIEW_SETTINGS.spectraLogPower = true;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectraLogFreq = true;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectraMinFreq = 0;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectraMaxFreq = 25;
+    DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectraPower = true;
+    DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectraPowerMemory = true;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectraMinPower = 1;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectraMaxPower = 5;
+    // spectrogram options
+    DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectrogramPower = false;
+    DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectrogramPowerMemory = true;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMinPower = 20;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMaxPower = 120;
     DEFAULT_WAVE_VIEW_SETTINGS.spectrogramOverlap = 0.859375;
-    DEFAULT_WAVE_VIEW_SETTINGS.minFreq = 0;
-    DEFAULT_WAVE_VIEW_SETTINGS.maxFreq = 25;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMinFreq = 0;
+    DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMaxFreq = 25;
     DEFAULT_WAVE_VIEW_SETTINGS.binSize = 2;
     DEFAULT_WAVE_VIEW_SETTINGS.nfft = 0; // Zero means automatic
+    DEFAULT_WAVE_VIEW_SETTINGS.spectrogramLogPower = true;
+    // particle motion options
+    DEFAULT_WAVE_VIEW_SETTINGS.useAlternateOrientationCode = false;
+    DEFAULT_WAVE_VIEW_SETTINGS.alternateOrientationCode = "Z12";
+    // filter otions
     DEFAULT_WAVE_VIEW_SETTINGS.filter = new Butterworth();
     DEFAULT_WAVE_VIEW_SETTINGS.filterOn = false;
     DEFAULT_WAVE_VIEW_SETTINGS.zeroPhaseShift = true;
-    DEFAULT_WAVE_VIEW_SETTINGS.useAlternateOrientationCode = false;
-    DEFAULT_WAVE_VIEW_SETTINGS.alternateOrientationCode = "Z12";
 
     List<String> candidateNames = new LinkedList<String>();
     candidateNames.add(DEFAULTS_FILENAME);
@@ -144,28 +173,47 @@ public class WaveViewSettings {
    */
   public void copy(WaveViewSettings s) {
     viewType = s.viewType;
-    removeBias = s.removeBias;
+    
+    // wave options
     autoScaleAmp = s.autoScaleAmp;
     autoScaleAmpMemory = s.autoScaleAmpMemory;
-    maxAmp = s.maxAmp;
-    minAmp = s.minAmp;
-    autoScalePowerMemory = s.autoScalePowerMemory;
-    autoScalePower = s.autoScalePower;
-    minPower = s.minPower;
-    maxPower = s.maxPower;
-    filter = new Butterworth(s.filter);
+    waveMaxAmp = s.waveMaxAmp;
+    waveMinAmp = s.waveMinAmp;
     useUnits = s.useUnits;
-    minFreq = s.minFreq;
-    maxFreq = s.maxFreq;
+    removeBias = s.removeBias;
+    
+    // spectra options
+    spectraLogPower = s.spectraLogPower;
+    spectraLogFreq = s.spectraLogFreq;
+    autoScaleSpectraPowerMemory = s.autoScaleSpectraPowerMemory;
+    autoScaleSpectraPower = s.autoScaleSpectraPower;
+    spectraMinPower = s.spectraMinPower;
+    spectraMaxPower = s.spectraMaxPower;
+    spectraMinFreq = s.spectraMinFreq;
+    spectraMaxFreq = s.spectraMaxFreq;
+    
+    // spectrogram options
+    autoScaleSpectrogramPowerMemory = s.autoScaleSpectrogramPowerMemory;
+    autoScaleSpectrogramPower = s.autoScaleSpectrogramPower;
+    spectrogramMinPower = s.spectrogramMinPower;
+    spectrogramMaxPower = s.spectrogramMaxPower;
+    spectrogramMinFreq = s.spectrogramMinFreq;
+    spectrogramMaxFreq = s.spectrogramMaxFreq;
     binSize = s.binSize;
     nfft = s.nfft;
     spectrogramOverlap = s.spectrogramOverlap;
-    logPower = s.logPower;
-    logFreq = s.logFreq;
-    zeroPhaseShift = s.zeroPhaseShift;
-    filterOn = s.filterOn;
+    spectrogramLogPower = s.spectrogramLogPower;
+    
+    // particle motion options
     useAlternateOrientationCode = s.useAlternateOrientationCode;
     alternateOrientationCode = s.alternateOrientationCode;
+    
+    // filter options
+    zeroPhaseShift = s.zeroPhaseShift;
+    filterOn = s.filterOn;
+    filter = new Butterworth(s.filter);
+    
+    // pick mode
     pickEnabled = s.pickEnabled;
   }
 
@@ -175,48 +223,88 @@ public class WaveViewSettings {
    */
   public void set(ConfigFile cf) {
     viewType = ViewType.fromString(cf.getString("viewType"));
-    filter.set(cf.getSubConfig("filter"));
-    maxAmp = StringUtils.stringToDouble(cf.getString("maxAmp"), DEFAULT_WAVE_VIEW_SETTINGS.maxAmp);
-    minAmp = StringUtils.stringToDouble(cf.getString("minAmp"), DEFAULT_WAVE_VIEW_SETTINGS.minAmp);
-    maxPower =
-        StringUtils.stringToDouble(cf.getString("maxPower"), DEFAULT_WAVE_VIEW_SETTINGS.maxPower);
-    minPower =
-        StringUtils.stringToDouble(cf.getString("minPower"), DEFAULT_WAVE_VIEW_SETTINGS.minPower);
-    minFreq =
-        StringUtils.stringToDouble(cf.getString("minFreq"), DEFAULT_WAVE_VIEW_SETTINGS.minFreq);
-    maxFreq =
-        StringUtils.stringToDouble(cf.getString("maxFreq"), DEFAULT_WAVE_VIEW_SETTINGS.maxFreq);
-    spectrogramOverlap = StringUtils.stringToDouble(cf.getString("spectrogramOverlap"),
-        DEFAULT_WAVE_VIEW_SETTINGS.spectrogramOverlap);
-
-    removeBias = StringUtils.stringToBoolean(cf.getString("removeBias"),
-        DEFAULT_WAVE_VIEW_SETTINGS.removeBias);
-    filterOn =
-        StringUtils.stringToBoolean(cf.getString("filterOn"), DEFAULT_WAVE_VIEW_SETTINGS.filterOn);
-    zeroPhaseShift = StringUtils.stringToBoolean(cf.getString("zeroPhaseShift"),
-        DEFAULT_WAVE_VIEW_SETTINGS.zeroPhaseShift);
+    
+    // wave settings
+    waveMaxAmp =
+        StringUtils.stringToDouble(cf.getString("waveMaxAmp"),
+            DEFAULT_WAVE_VIEW_SETTINGS.waveMaxAmp);
+    waveMinAmp =
+        StringUtils.stringToDouble(cf.getString("waveMinAmp"),
+            DEFAULT_WAVE_VIEW_SETTINGS.waveMinAmp);
     autoScaleAmp = StringUtils.stringToBoolean(cf.getString("autoScaleAmp"),
         DEFAULT_WAVE_VIEW_SETTINGS.autoScaleAmp);
     autoScaleAmpMemory = StringUtils.stringToBoolean(cf.getString("autoScaleAmpMemory"),
         DEFAULT_WAVE_VIEW_SETTINGS.autoScaleAmpMemory);
-    autoScalePower = StringUtils.stringToBoolean(cf.getString("autoScalePower"),
-        DEFAULT_WAVE_VIEW_SETTINGS.autoScalePower);
-    autoScalePowerMemory = StringUtils.stringToBoolean(cf.getString("autoScalePowerMemory"),
-        DEFAULT_WAVE_VIEW_SETTINGS.autoScalePowerMemory);
+    removeBias = StringUtils.stringToBoolean(cf.getString("removeBias"),
+        DEFAULT_WAVE_VIEW_SETTINGS.removeBias);
     useUnits =
         StringUtils.stringToBoolean(cf.getString("useUnits"), DEFAULT_WAVE_VIEW_SETTINGS.useUnits);
-    logFreq =
-        StringUtils.stringToBoolean(cf.getString("logFreq"), DEFAULT_WAVE_VIEW_SETTINGS.logFreq);
-    logPower =
-        StringUtils.stringToBoolean(cf.getString("logPower"), DEFAULT_WAVE_VIEW_SETTINGS.logPower);
+    
+    // spectra settings
+    spectraLogFreq =
+        StringUtils.stringToBoolean(cf.getString("spectraLogFreq"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectraLogFreq);
+    spectraLogPower =
+        StringUtils.stringToBoolean(cf.getString("spectraLogPower"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectraLogPower);
+    spectraMinFreq =
+        StringUtils.stringToDouble(cf.getString("spectraMinFreq"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectraMinFreq);
+    spectraMaxFreq =
+        StringUtils.stringToDouble(cf.getString("spectraMaxFreq"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectraMaxFreq);
+    autoScaleSpectraPower = StringUtils.stringToBoolean(cf.getString("autoScaleSpectraPower"),
+        DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectraPower);
+    autoScaleSpectraPowerMemory =
+        StringUtils.stringToBoolean(cf.getString("autoScaleSpectraPowerMemory"),
+            DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectraPowerMemory);
+    spectraMaxPower =
+        StringUtils.stringToDouble(cf.getString("spectraMaxPower"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectraMaxPower);
+    spectraMinPower =
+        StringUtils.stringToDouble(cf.getString("spectraMinPower"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectraMinPower);
+    
+    // spectrogram settings
+    spectrogramMinFreq =
+        StringUtils.stringToDouble(cf.getString("spectrogramMinFreq"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMinFreq);
+    spectrogramMaxFreq =
+        StringUtils.stringToDouble(cf.getString("spectrogramMaxFreq"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMaxFreq);
+    autoScaleSpectrogramPower = StringUtils.stringToBoolean(cf.getString("autoScaleSpectrogramPower"),
+        DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectrogramPower);
+    autoScaleSpectrogramPowerMemory =
+        StringUtils.stringToBoolean(cf.getString("autoScaleSpectrogramPowerMemory"),
+            DEFAULT_WAVE_VIEW_SETTINGS.autoScaleSpectrogramPowerMemory);
+    spectrogramMaxPower =
+        StringUtils.stringToDouble(cf.getString("spectrogramMaxPower"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMaxPower);
+    spectrogramMinPower =
+        StringUtils.stringToDouble(cf.getString("spectrogramMinPower"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectrogramMinPower);
     binSize =
         StringUtils.stringToDouble(cf.getString("binSize"), DEFAULT_WAVE_VIEW_SETTINGS.binSize);
     nfft = StringUtils.stringToInt(cf.getString("nfft"), DEFAULT_WAVE_VIEW_SETTINGS.nfft);
+    spectrogramOverlap = StringUtils.stringToDouble(cf.getString("spectrogramOverlap"),
+        DEFAULT_WAVE_VIEW_SETTINGS.spectrogramOverlap);
+    spectrogramLogPower =
+        StringUtils.stringToBoolean(cf.getString("spectrogramLogPower"),
+            DEFAULT_WAVE_VIEW_SETTINGS.spectrogramLogPower);
+    
+    // particle motion settings
     useAlternateOrientationCode =
         StringUtils.stringToBoolean(cf.getString("useAlternateOrientationCode"),
             DEFAULT_WAVE_VIEW_SETTINGS.useAlternateOrientationCode);
     alternateOrientationCode = StringUtils.stringToString(cf.getString("alternateOrientationCode"),
         DEFAULT_WAVE_VIEW_SETTINGS.alternateOrientationCode);
+    
+    // filter settings
+    filterOn =
+        StringUtils.stringToBoolean(cf.getString("filterOn"), DEFAULT_WAVE_VIEW_SETTINGS.filterOn);
+    filter.set(cf.getSubConfig("filter"));
+    zeroPhaseShift = StringUtils.stringToBoolean(cf.getString("zeroPhaseShift"),
+        DEFAULT_WAVE_VIEW_SETTINGS.zeroPhaseShift);
   }
 
   /**
@@ -227,27 +315,39 @@ public class WaveViewSettings {
   public void save(ConfigFile cf, String prefix) {
     cf.put(prefix + ".viewType", viewType.code);
     filter.save(cf, prefix + ".filter");
-    cf.put(prefix + ".maxAmp", Double.toString(maxAmp));
-    cf.put(prefix + ".minAmp", Double.toString(minAmp));
-    cf.put(prefix + ".minPower", Double.toString(minPower));
-    cf.put(prefix + ".maxPower", Double.toString(maxPower));
-    cf.put(prefix + ".minFreq", Double.toString(minFreq));
-    cf.put(prefix + ".maxFreq", Double.toString(maxFreq));
-    cf.put(prefix + ".spectrogramOverlap", Double.toString(spectrogramOverlap));
-    cf.put(prefix + ".removeBias", Boolean.toString(removeBias));
-    cf.put(prefix + ".filterOn", Boolean.toString(filterOn));
-    cf.put(prefix + ".zeroPhaseShift", Boolean.toString(zeroPhaseShift));
+    // wave settings
     cf.put(prefix + ".autoScaleAmp", Boolean.toString(autoScaleAmp));
     cf.put(prefix + ".autoScaleAmpMemory", Boolean.toString(autoScaleAmpMemory));
-    cf.put(prefix + ".autoScalePower", Boolean.toString(autoScalePower));
-    cf.put(prefix + ".autoScalePowerMemory", Boolean.toString(autoScalePowerMemory));
+    cf.put(prefix + ".waveMaxAmp", Double.toString(waveMaxAmp));
+    cf.put(prefix + ".waveMinAmp", Double.toString(waveMinAmp));
+    cf.put(prefix + ".removeBias", Boolean.toString(removeBias));
     cf.put(prefix + ".useUnits", Boolean.toString(useUnits));
-    cf.put(prefix + ".logFreq", Boolean.toString(logFreq));
-    cf.put(prefix + ".logPower", Boolean.toString(logPower));
+    // spectra settings
+    cf.put(prefix + ".spectraLogFreq", Boolean.toString(spectraLogFreq));
+    cf.put(prefix + ".spectraLogPower", Boolean.toString(spectraLogPower));
+    cf.put(prefix + ".spectraMinFreq", Double.toString(spectraMinFreq));
+    cf.put(prefix + ".spectraMaxFreq", Double.toString(spectraMaxFreq));
+    cf.put(prefix + ".autoScaleSpectrogramPower", Boolean.toString(autoScaleSpectrogramPower));
+    cf.put(prefix + ".autoScaleSpectrogramPowerMemory", Boolean.toString(autoScaleSpectrogramPowerMemory));
+    cf.put(prefix + ".spectraMinPower", Double.toString(spectraMinPower));
+    cf.put(prefix + ".spectraMaxPower", Double.toString(spectraMaxPower));
+    // spectrogram settings
+    cf.put(prefix + ".spectrogramMinFreq", Double.toString(spectrogramMinFreq));
+    cf.put(prefix + ".spectrogramMaxFreq", Double.toString(spectrogramMaxFreq));
+    cf.put(prefix + ".autoScalePower", Boolean.toString(autoScaleSpectrogramPower));
+    cf.put(prefix + ".autoScalePowerMemory", Boolean.toString(autoScaleSpectrogramPowerMemory));
+    cf.put(prefix + ".spectrogramMinPower", Double.toString(spectrogramMinPower));
+    cf.put(prefix + ".spectrogramMaxPower", Double.toString(spectrogramMaxPower));
     cf.put(prefix + ".binSize", Double.toString(binSize));
     cf.put(prefix + ".nfft", Integer.toString(nfft));
+    cf.put(prefix + ".spectrogramOverlap", Double.toString(spectrogramOverlap));
+    cf.put(prefix + ".spectrogramLogPower", Boolean.toString(spectrogramLogPower));
+    // particle motion settings
     cf.put(prefix + ".useAlternateOrientationCode", Boolean.toString(useAlternateOrientationCode));
     cf.put(prefix + ".alternateOrientationCode", alternateOrientationCode);
+    // filter settings
+    cf.put(prefix + ".filterOn", Boolean.toString(filterOn));
+    cf.put(prefix + ".zeroPhaseShift", Boolean.toString(zeroPhaseShift));
   }
 
   public void setType(ViewType t) {
@@ -283,22 +383,27 @@ public class WaveViewSettings {
    */
   public void cycleLogSettings() {
 
-    if (logFreq == logPower) {
-      logPower = !logPower;
+    if (spectraLogFreq == spectraLogPower) {
+      spectraLogPower = !spectraLogPower;
     } else {
-      logFreq = !logFreq;
+      spectraLogFreq = !spectraLogFreq;
     }
 
     notifyView();
   }
 
-  public void toggleLogFreq() {
-    logFreq = !logFreq;
+  public void toggleSpectraLogFreq() {
+    spectraLogFreq = !spectraLogFreq;
     notifyView();
   }
 
-  public void toggleLogPower() {
-    logPower = !logPower;
+  public void toggleSpectraLogPower() {
+    spectraLogPower = !spectraLogPower;
+    notifyView();
+  }
+  
+  public void toggleSpectrogramLogPower() {
+    spectrogramLogPower = !spectrogramLogPower;
     notifyView();
   }